revert to FabricVLANs test
