+++ ui/src/app/subnets/views/FabricDetails/FabricDetails.test.tsx
@@ -1,4 +1,4 @@
-import { render, screen } from "@testing-library/react";
+import { render, screen, waitFor } from "@testing-library/react";
 import { Provider } from "react-redux";
 import { MemoryRouter, Route } from "react-router-dom";
 import configureStore from "redux-mock-store";
@@ -10,13 +10,6 @@ import subnetsURLs from "app/subnets/urls";
 import {
   fabricState as fabricStateFactory,
   rootState as rootStateFactory,
-  fabric as fabricFactory,
-  vlan as vlanFactory,
-  vlanState as vlanStateFactory,
-  space as spaceFactory,
-  spaceState as spaceStateFactory,
-  subnet as subnetFactory,
-  subnetState as subnetStateFactory,
 } from "testing/factories";
 
 const mockStore = configureStore();
@@ -134,66 +127,5 @@ it("shows a spinner if the fabric has not loaded yet", () => {
 
   expect(
     screen.getByTestId("section-header-title-spinner")
-  ).toBeInTheDocument();
-});
-
-it("renders correct details", () => {
-  const fabric = fabricFactory({ id: 1 });
-  const state = rootStateFactory({
-    vlan: vlanStateFactory({
-      loaded: true,
-      loading: false,
-      items: [vlanFactory()],
-    }),
-    space: spaceStateFactory({
-      loaded: true,
-      loading: false,
-      items: [spaceFactory()],
-    }),
-    subnet: subnetStateFactory({
-      loaded: true,
-      loading: false,
-      items: [subnetFactory()],
-    }),
-    fabric: fabricStateFactory({
-      items: [],
-      loaded: true,
-      loading: false,
-    }),
-  });
-  const store = mockStore(state);
-
-  render(
-    <Provider store={store}>
-      <MemoryRouter
-        initialEntries={[{ pathname: subnetsURLs.fabric.index({ id: 1 }) }]}
-      >
-        <Route
-          exact
-          path={subnetsURLs.space.index({ id: fabric.id })}
-          component={() => <FabricDetails />}
-        />
-      </MemoryRouter>
-    </Provider>
-  );
-
-  expect(
-    screen.getByRole("heading", { name: "VLANs on this fabric" })
-  ).toBeInTheDocument();
-
-  expect(
-    screen.getByRole("columnheader", { name: "VLAN" })
-  ).toBeInTheDocument();
-
-  expect(
-    screen.getByRole("columnheader", { name: "Subnet" })
-  ).toBeInTheDocument();
-
-  expect(
-    screen.getByRole("columnheader", { name: "Available" })
-  ).toBeInTheDocument();
-
-  expect(
-    screen.getByRole("columnheader", { name: "Space" })
   ).toBeInTheDocument();
 });
